rename ts file
