+++ model/base/factory.ts
@@ -1,7 +1,7 @@
 import type { Brand, Expand } from 'devtypes/types/util';
 import type { EntityType, RegistryType } from '../../enum/registry/system';
 import type { Collection } from './modifier';
-import type { Struct } from './primitives';
+import type { Struct } from './primitive';
 
 export type Factory<
   T extends EntityType | RegistryType,
