+++ model/base/primitive.ts
