+++ model/base/value.ts
@@ -3,7 +3,7 @@ import type { Primitive } from 'devtypes/types/primitive';
 import type { Brand, Expand } from 'devtypes/types/util';
 import type { ValueOrigin, ValueType } from '../../enum/base/value';
 import type { NoUnit, PhysicalQuantity, UnitId } from '../registry/unit';
-import type { Struct } from './primitives';
+import type { Struct } from './primitive';
 import type { Uncertainty } from './uncertainty';
 
 type UnitField< Q extends PhysicalQuantity | NoUnit > = [ Q ] extends [ NoUnit ] ? {} : {
