+++ model/collection/generic.ts
@@ -2,7 +2,7 @@ import type { RequireAtLeastOne } from 'devtypes/types/constraint';
 import type { Currency } from '../../enum/base/locale';
 import type { PriceContext, PriceTax, PriceValue } from '../../enum/collection/generic';
 import type { Collection, Many } from '../base/modifier';
-import type { IsoDate } from '../base/primitives';
+import type { IsoDate } from '../base/primitive';
 import type { StructProperty } from '../base/property';
 import type { NumberValue } from '../base/value';
 import type { PhysicalQuantity } from '../registry/unit';
